+++ apps/angular-demo/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, inject, Injector } from '@angular/core';
 import {
   ClipboardPastedEvent,
   configureShortcuts,
@@ -25,6 +25,7 @@ import {
   type Node,
   type Port,
 } from 'ng-diagram';
+import { defaultModel } from './data/default-model';
 import { nodeTemplateMap } from './data/node-template';
 import { paletteModel } from './data/palette-model';
 import { ButtonEdgeComponent } from './edge-template/button-edge/button-edge.component';
@@ -43,6 +44,8 @@ import { ToolbarComponent } from './toolbar/toolbar.component';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {
+  private readonly injector = inject(Injector);
+
   paletteModel: NgDiagramPaletteItem[] = paletteModel;
   nodeTemplateMap: NgDiagramNodeTemplateMap = nodeTemplateMap;
   edgeTemplateMap = new NgDiagramEdgeTemplateMap([
@@ -197,253 +200,9 @@ export class AppComponent {
     });
   }
 
-  model = initializeModel({
-    nodes: [
-      {
-        id: '1',
-        type: 'image',
-        position: { x: 100, y: 50 },
-        data: { imageUrl: 'https://tinyurl.com/bddnt44s' },
-      },
-      { id: '2', type: 'input-field', position: { x: 400, y: 100 }, data: {} },
-      {
-        id: '3',
-        type: 'resizable',
-        position: { x: 700, y: 200 },
-        size: { width: 300, height: 400 },
-        autoSize: false,
-        data: {},
-        angle: 45,
-      },
-      {
-        id: '4',
-        position: { x: 800, y: 350 },
-        data: {},
-        isGroup: true,
-      },
-      {
-        id: '5',
-        position: { x: 100, y: 250 },
-        data: { label: 'edge is manual' },
-      },
-      {
-        id: '6',
-        position: { x: 463, y: 382 },
-        data: { label: "so it's ok it's unconnected after move" },
-      },
-      {
-        id: '7',
-        position: { x: 100, y: 450 },
-        data: {},
-      },
-      {
-        id: '8',
-        position: { x: 550, y: 550 },
-        data: {},
-      },
-      {
-        id: '9',
-        position: { x: 100, y: 650 },
-        data: { label: 'just bezier edge' },
-      },
-      {
-        id: '10',
-        position: { x: 450, y: 750 },
-        data: {},
-      },
-      {
-        id: '11',
-        position: { x: 600, y: 650 },
-        data: { label: 'test linking' },
-      },
-      {
-        id: '12',
-        position: { x: 1000, y: 550 },
-        data: {},
-      },
-      {
-        id: '12',
-        position: { x: 700, y: 750 },
-        data: {},
-        type: 'customized-default',
-        resizable: true,
-        rotatable: true,
-      },
-      {
-        id: '13',
-        position: { x: 1200, y: 250 },
-        data: { text: 'SN 777888' },
-        type: 'chip',
-        autoSize: false,
-        size: { width: 200, height: 200 },
-      },
-      {
-        id: '14',
-        position: { x: 1200, y: 550 },
-        data: { text: 'SN 877889' },
-        type: 'chip',
-        autoSize: false,
-        size: { width: 200, height: 300 },
-      },
-      {
-        id: '15',
-        position: { x: -200, y: 500 },
-        data: {},
-        size: { width: 300, height: 200 },
-        autoSize: false,
-        type: 'custom-group',
-        isGroup: true,
-        rotatable: true,
-      },
-    ],
-    edges: [
-      {
-        id: '1',
-        source: '1',
-        target: '2',
-        data: {},
-        sourcePort: 'port-right',
-        targetPort: 'port-left',
-        routing: 'orthogonal',
-      },
-      {
-        id: '2',
-        source: '2',
-        target: '3',
-        data: {},
-        sourcePort: 'port-right',
-        targetPort: 'port-left-1',
-        type: 'button-edge',
-      },
-      {
-        id: '3',
-        source: '5',
-        target: '6',
-        data: { labelPosition: '0.45' },
-        sourcePort: 'port-right',
-        targetPort: 'port-left',
-        type: 'labelled-edge',
-        routing: 'orthogonal',
-        routingMode: 'manual',
-        points: [
-          { x: 300, y: 274 },
-          { x: 380, y: 274 },
-          { x: 380, y: 314 },
-          { x: 420, y: 314 },
-          { x: 420, y: 354 },
-          { x: 380, y: 354 },
-          { x: 380, y: 407 },
-          { x: 460, y: 407 },
-        ],
-      },
-      {
-        id: '4',
-        source: '7',
-        target: '8',
-        data: {},
-        sourcePort: 'port-right',
-        targetPort: 'port-left',
-        type: 'custom-polyline-edge',
-      },
-      {
-        id: '5',
-        source: '9',
-        target: '10',
-        data: { labelPosition: 0.7 },
-        type: 'labelled-edge',
-        routing: 'bezier',
-      },
-      {
-        id: '6',
-        source: '11',
-        target: '12',
-        data: {},
-        sourcePort: 'port-right',
-        targetPort: 'port-left',
-        type: 'dashed-edge',
-        routing: 'orthogonal',
-      },
-      {
-        id: '11',
-        data: {},
-        source: '13',
-        sourcePort: 'port-left-4',
-        target: '14',
-        targetPort: 'port-left-1',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '12',
-        data: {},
-        source: '13',
-        sourcePort: 'port-left-4',
-        target: '14',
-        targetPort: 'port-left-2',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '13',
-        data: {},
-        source: '13',
-        sourcePort: 'port-left-4',
-        target: '14',
-        targetPort: 'port-left-3',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '14',
-        data: {},
-        source: '14',
-        sourcePort: 'port-right-3',
-        target: '13',
-        targetPort: 'port-right-4',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '15',
-        data: {},
-        source: '13',
-        sourcePort: 'port-right-3',
-        target: '14',
-        targetPort: 'port-right-4',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '16',
-        data: {},
-        source: '13',
-        sourcePort: 'port-right-2',
-        target: '14',
-        targetPort: 'port-right-1',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '17',
-        data: {},
-        source: '14',
-        sourcePort: 'port-right-2',
-        target: '13',
-        targetPort: 'port-right-2',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '18',
-        data: {},
-        source: '13',
-        sourcePort: 'port-right-1',
-        target: '14',
-        targetPort: 'port-left-4',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-      {
-        id: '19',
-        data: {},
-        source: '13',
-        sourcePort: 'port-left-2',
-        target: '14',
-        targetPort: 'port-right-1',
-        targetArrowhead: 'ng-diagram-arrow',
-      },
-    ],
-  });
+  onReinitializeModel(): void {
+    this.model = initializeModel(defaultModel, this.injector);
+  }
+
+  model = initializeModel(defaultModel);
 }