+++ packages/ng-diagram/projects/ng-diagram/src/core/src/flow-core.test.ts
@@ -207,6 +207,7 @@ describe('FlowCore', () => {
         mockEnvironment,
         undefined,
         undefined,
+        undefined,
         customConfig
       );